Control global de excepciones

+++ src/service/TaskService.py
@@ -1,8 +1,11 @@
-from flask import request, jsonify
+from urllib.error import HTTPError
+
+from flask import request, Response
 
 from ..config.database import db
 from ..models.TaskModel import Task
 from ..schemas.TaskSchema import task_schema, tasks_schema
+from ..util.Constants import RECORD_NOT_FOUND
 
 
 def create_task():
@@ -11,13 +14,13 @@ def create_task():
     new_task = Task(title, description)
     db.session.add(new_task)
     db.session.commit()
-    return task_schema.jsonify(new_task)
+    return Response(task_schema.jsonify(new_task), 201, mimetype='application/json')
 
 
 def get_tasks():
     all_tasks = Task.query.all()
     result = tasks_schema.dump(all_tasks)
-    return jsonify(result)
+    return Response(result)
 
 
 def get_task(id):
@@ -35,6 +38,8 @@ def update_task(id):
 
 def delete_task(id):
     task = Task.query.get(id)
+    if task is None:
+        raise HTTPError(request.url, 404, RECORD_NOT_FOUND.format(id), None, None)
     db.session.delete(task)
     db.session.commit()
     return task_schema.jsonify(task)
